requirements.txt
playwright
playwright_stealth
requests
pyyaml
pydantic
dotenv